+++ android/src/main/kotlin/io/aban/admob_nextgen/core/AdErrorExt.kt
@@ -1,5 +1,6 @@
 package io.aban.admob_nextgen.core
 
+import com.google.android.libraries.ads.mobile.sdk.common.AdInspectorError
 import com.google.android.libraries.ads.mobile.sdk.common.FullScreenContentError
 import com.google.android.libraries.ads.mobile.sdk.common.LoadAdError
 
@@ -13,4 +14,7 @@ internal fun LoadAdError.toFlutterMap(): Map<String, Any> =
     codecSafeErrorMap(code.value, message)
 
 internal fun FullScreenContentError.toFlutterMap(): Map<String, Any> =
+    codecSafeErrorMap(code.value, message)
+
+internal fun AdInspectorError.toFlutterMap(): Map<String, Any> =
     codecSafeErrorMap(code.value, message)